Deprecate topitem(...)
Some people keep using topitem(...), and it's bad, because there is no "universal topitem". It can be a different item depending on the action you are trying to do.

i.e.: on a rope spot, if you useitemon(...) it, then the top item will be the rope tile... If you moveitems(...) on it, then the top item will be the item or creature above it.